Smarty now upgraded from version 2 to version 3.1.8
NB: Smarty 3 has syntax changes and is a complete rewrite. This has necessitated changing a lot of method calls in ScriptMind::Links internals 

The translation class IntSmarty is known not to work properly with Smarty 3. I've made some changes and it seems to be working, but I can't guarantee that it is thoroughly tested. IntSmarty is now optional and is only used if config.php contains the line
define( 'USE_INTSMARTY', 1 );

For now not using IntSmarty is the default. This disables translations.

Resolves issue #2